requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
